+++ ShiftSchedularEntity/Models/DataTransferObjects/Incoming/AddEntityRuleSpecificationDTO.cs
@@ -4,7 +4,7 @@
     {
         public Guid EntityRuleId { get; set; }
         public int SpecificationId { get; set; }
-        public int RuleSpecificationValue { get; set; }
+        public float RuleSpecificationValue { get; set; }
         public string AspectReferenceId { get; set; }
         public int BusinessAspectId { get; set; }
         public string AspectReferenceId2 { get; set; }
